requirements.txt
python-binance==1.0.19
python-dotenv==1.0.0
pandas==2.1.4
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==3.0.1
cryptography==41.0.7
gunicorn==21.2.0
